# packages/api/src/generated/models/resendVerificationRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Broker API
 * OpenAPI spec version: 0.1.0
 */
import type { ResendVerificationRequestClientApp } from './resendVerificationRequestClientApp';

export interface ResendVerificationRequest {
  email: string;
  client_app: ResendVerificationRequestClientApp;
}