# packages/api/src/generated/models/userRegister.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Broker API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRegisterClientApp } from './userRegisterClientApp';

export interface UserRegister {
  /**
   * @minLength 1
   * @maxLength 255
   */
  name: string;
  email: string;
  /**
   * @minLength 8
   * @maxLength 128
   */
  password: string;
  client_app: UserRegisterClientApp;
}